changed id type to int(eger)

+++ java_mysql_curd_api/src/main/java/com/backend/java_mysql_curd_api/service/StudentService.java
@@ -21,7 +21,7 @@ public class StudentService {
     }
 
     // FETCHING STUDENT BY ID
-    public Optional<Student> getStudentById(Long student_id){
+    public Optional<Student> getStudentById(int student_id){
         return studentRepository.findById(student_id);
     }
